fix: move SimCfnResolveContext to separate file
