+++ src/service/cloudformation/template/node/fn/get-att/sim-cfn-fn-get-att.ts
@@ -1,5 +1,6 @@
-import { SimCfnNode, type SimCfnResolveContext } from "../../sim-cfn-node.js";
+import { SimCfnNode } from "../../sim-cfn-node.js";
 import type { SimCfnTemplateValue } from "../../../value/sim-cfn-template-value.js";
+import type { SimCfnResolveContext } from "../../../resolve/sim-cfn-resolve-context.js";
 
 /**
  * A CloudFormation `{ "Fn::GetAtt": ["LogicalId", "AttributeName"] }`
