+++ src/service/cloudformation/template/node/sim-cfn-node.ts
@@ -1,18 +1,5 @@
-import type { SimCfnParameters } from "../../parameters/sim-cfn-parameters.js";
 import type { SimCfnTemplateValue } from "../value/sim-cfn-template-value.js";
-import type { SimCfnResourceRefResolver } from "../resolve/sim-cfn-resource-ref-resolver.js";
-import type { SimCfnPseudoParameters } from "../../parameters/pseudo/sim-cfn-pseudo-parameters.js";
-
-/**
- * Context available while resolving a parsed CloudFormation node tree.
- */
-export class SimCfnResolveContext {
-  constructor(
-    readonly parameters: SimCfnParameters,
-    readonly resources?: SimCfnResourceRefResolver | undefined,
-    readonly pseudoParameters?: SimCfnPseudoParameters | undefined,
-  ) {}
-}
+import type { SimCfnResolveContext } from "../resolve/sim-cfn-resolve-context.js";
 
 /**
  * A parsed CloudFormation template value.
