+++ src/service/cloudformation/template/node/sim-cfn-ref.ts
@@ -1,5 +1,6 @@
-import { SimCfnNode, type SimCfnResolveContext } from "./sim-cfn-node.js";
+import { SimCfnNode } from "./sim-cfn-node.js";
 import type { SimCfnTemplateValue } from "../value/sim-cfn-template-value.js";
+import type { SimCfnResolveContext } from "../resolve/sim-cfn-resolve-context.js";
 
 /**
  * A CloudFormation `{ "Ref": "Name" }` expression.
