+++ src/service/cloudformation/template/parse/node/sim-cfn-node-parser.iso.test.ts
@@ -11,11 +11,12 @@ import { SimCfnParameters } from "../../../parameters/sim-cfn-parameters.js";
 import { SimCfnFnJoin } from "../../node/fn/join/sim-cfn-fn-join.js";
 import { SimCfnList } from "../../node/sim-cfn-list.js";
 import { SimCfnLiteral } from "../../node/sim-cfn-literal.js";
-import { SimCfnNode, SimCfnResolveContext } from "../../node/sim-cfn-node.js";
+import { SimCfnNode } from "../../node/sim-cfn-node.js";
 import { SimCfnObject } from "../../node/sim-cfn-object.js";
 import { SimCfnRef } from "../../node/sim-cfn-ref.js";
 import { parseSimCfnNode, SimCfnNodeParser } from "./sim-cfn-node-parser.js";
 import { SimAws } from "../../../../aws/sim-aws.js";
+import { SimCfnResolveContext } from "../../resolve/sim-cfn-resolve-context.js";
 
 describe("SimCfnNodeParser", () => {
   it("parses arrays into CloudFormation list nodes recursively", () => {
